+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Release History
 
-## 0.18.1 (unreleased)
+## 0.18.1 (Unreleased)
 
 ### Other Changes
 
 * The connection mux goroutine has been removed, eliminating a potential source of deadlocks.
 * Automatic link flow control is built on the manual creditor.
+* Clarified docs that messages received from a sender configured in a mode other than `SenderSettleModeSettled` must be acknowledged.
 
 ## 0.18.0 (2022-12-06)